% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_utils.R
\name{addSpecies}
\alias{addSpecies}
\alias{addSpecies.MRMizerParams}
\title{Add species to an MRMizerParams model}
\usage{
\method{addSpecies}{MRMizerParams}(
  params,
  species_params,
  interaction,
  gear_params = data.frame(),
  initial_effort,
  resource_interaction = NULL,
  info_level = 0,
  ...
)
}
\arguments{
\item{params}{An \linkS4class{MRMizerParams} object.}

\item{species_params}{A data frame with the species parameters of the new
species.}

\item{interaction}{Interaction matrix. A square matrix giving either the
interaction coefficients between all species or only those between the new
species. In the latter case all interaction between an old and a new
species are set to 1. If this argument is missing, all interactions
involving a new species are set to 1.}

\item{gear_params}{Data frame with the gear parameters for the new
species. If not provided then the new species will not be fished.}

\item{initial_effort}{A named vector with the effort for any new fishing gear
introduced in \code{gear_params}. Not needed if the added species are only
fished by already existing gear. Should not include effort values
for existing gear. New gear for which no effort is set via this
vector will have an initial effort of 0.}

\item{resource_interaction}{Optional matrix (new species x resources) of
interaction values between new species and each resource. Defaults to 1
for all new species and resources.}

\item{info_level}{Controls the amount of information messages that are shown
when the function sets default values for parameters. Higher levels lead
to more messages. Set to 0 to suppress all such messages.}

\item{...}{Additional arguments}
}
\value{
An \linkS4class{MRMizerParams} object with the new species added.
}
\description{
Extends \code{\link[mizer:addSpecies]{mizer::addSpecies()}} to handle the multiple-resource interaction
matrix. New species are added with a resource interaction of 1 for all
resources by default, or as specified by \code{resource_interaction}.
}
