% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMultipleResources.R
\name{balanceComponent}
\alias{balanceComponent}
\title{Replace the rate or the capacity by their balanced values}
\usage{
balanceComponent(params, derive_rate)
}
\arguments{
\item{params}{A MizerParams object with multiple resources set up.}

\item{derive_rate}{If TRUE the rate is calculated from the capacity,
otherwise the capacity is calculated from the rate.}
}
\value{
The updated MizerParams object.
}
\description{
Calls \code{\link[=balanceResources]{balanceResources()}} and stores the result in the MR component. The
derived array is marked with a comment so that it is not overwritten by the
values calculated from the resource parameters at the next call to
\code{\link[=setMultipleResources]{setMultipleResources()}}.
}
\keyword{internal}
