% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerMR-package.R
\name{.onLoad}
\alias{.onLoad}
\title{Register mizerMR with mizer}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library path supplied by R.}

\item{pkgname}{Package name supplied by R.}
}
\value{
Called for its side effect.
}
\description{
Registers the package as a mizer extension when the namespace is loaded.
}
\keyword{internal}
