% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_utils.R
\name{expandSizeGrid}
\alias{expandSizeGrid}
\alias{expandSizeGrid.MRMizerParams}
\title{Expand the size grid of an MRMizerParams model}
\usage{
\method{expandSizeGrid}{MRMizerParams}(
  params,
  new_min_w = min(params@w),
  new_max_w = max(params@w),
  preserve_species = params@species_params$species
)
}
\arguments{
\item{params}{An \linkS4class{MRMizerParams} object.}

\item{new_min_w}{The new minimum size in the grid. Must not be larger than
the current minimum size.}

\item{new_max_w}{The new maximum size in the grid. Must not be smaller than
the current maximum size.}

\item{preserve_species}{A vector of species names for which all rate arrays
should be copied over to the new params object rather than being
re-calculated from the species parameters. If missing, all species are
preserved.}
}
\value{
An \linkS4class{MRMizerParams} object with an expanded size grid.
}
\description{
Extends \code{\link[mizer:expandSizeGrid]{mizer::expandSizeGrid()}} to handle \linkS4class{MRMizerParams}
objects. The resource rate, capacity, and initial abundance arrays are
recalculated on the new grid from the resource parameters.
}
