% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{expect_unchanged}
\alias{expect_unchanged}
\title{Expect a mizer object to be unchanged}
\usage{
expect_unchanged(object, expected)
}
\arguments{
\item{object}{A \linkS4class{MizerParams} or \linkS4class{MizerSim} object.}

\item{expected}{The expected object.}
}
\value{
The expectation result from \code{\link[testthat:expect_equal]{testthat::expect_equal()}}.
}
\description{
Compares two mizer objects after normalising fields that are expected to
change during setup or projection.
}
\keyword{internal}
