% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{getResourceMort.mizerMRResourceView}
\alias{getResourceMort.mizerMRResourceView}
\title{Resource mortality of a single-resource view}
\usage{
\method{getResourceMort}{mizerMRResourceView}(
  params,
  n = initialN(params),
  n_pp = params@initial_n_pp,
  n_other = initialNOther(params),
  t = 0,
  ...
)
}
\arguments{
\item{params}{A \code{mizerMRResourceView} object.}

\item{n}{Unused.}

\item{n_pp}{Unused.}

\item{n_other}{Unused.}

\item{t}{Unused.}

\item{...}{Unused.}
}
\value{
The stored mortality vector.
}
\description{
Returns the mortality that was stored in the view when it was created, see
\link[=mizerMRResourceView-class]{mizerMRResourceView}.
}
\keyword{internal}
