% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayResource-class.R
\name{is.MRArrayResourceBySize}
\alias{is.MRArrayResourceBySize}
\title{Test if an object is an MRArrayResourceBySize}
\usage{
is.MRArrayResourceBySize(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{MRArrayResourceBySize} object, \code{FALSE} otherwise.
}
\description{
Test if an object is an MRArrayResourceBySize
}
