% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayResource-class.R
\name{is.MRArrayTimeByResourceBySize}
\alias{is.MRArrayTimeByResourceBySize}
\title{Test if an object is an MRArrayTimeByResourceBySize}
\usage{
is.MRArrayTimeByResourceBySize(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{MRArrayTimeByResourceBySize} object, \code{FALSE}
otherwise.
}
\description{
Test if an object is an MRArrayTimeByResourceBySize
}
