% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizer_internals.R
\name{mizer_fn}
\alias{mizer_fn}
\title{Get an unexported function from the mizer namespace}
\usage{
mizer_fn(name)
}
\arguments{
\item{name}{The name of the mizer function.}
}
\value{
The function.
}
\description{
Get an unexported function from the mizer namespace
}
\keyword{internal}
