% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerMR-class.R
\name{mizerMR-class}
\alias{mizerMR-class}
\alias{mizerMRSim-class}
\title{mizerMR marker classes}
\description{
S3 marker classes for \link[mizer:MizerParams]{MizerParams} and \link[mizer:MizerSim]{MizerSim} that enable S3 dispatch
for MR-specific methods.
}
\details{
Objects of class \code{mizerMR} are created by \code{\link[=setMultipleResources]{setMultipleResources()}}.
Objects of class \code{mizerMRSim} are returned automatically by \code{\link[mizer:project]{project()}}
when called on a \code{mizerMR} params object.

The classes are managed by mizer when the package is loaded: \code{.onLoad()} calls
\code{\link[mizer:registerExtension]{mizer::registerExtension()}}, which recognises mizerMR as a dispatching
extension from the S3 methods it registers for its marker class and inserts
\code{mizerMR} into the S3 class vector in dispatch order relative to any other
extension packages loaded in the same session.
}
\keyword{internal}
