% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerMR-package.R
\docType{package}
\name{mizerMR-package}
\alias{mizerMR}
\alias{mizerMR-package}
\title{mizerMR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sizespectrum/mizerMR}
  \item Report bugs at \url{https://github.com/sizespectrum/mizerMR/issues}
}

}
\author{
\strong{Maintainer}: Gustav Delius \email{gustav.delius@york.ac.uk} (\href{https://orcid.org/0000-0003-4092-8228}{ORCID}) [copyright holder]

}
\keyword{internal}
