% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_dynamics.R
\name{mizerMR_dynamics}
\alias{mizerMR_dynamics}
\title{Multiple-resource dynamics}
\usage{
mizerMR_dynamics(params, n_other, n_pp, rates, ...)
}
\description{
Internal rate hook registered by \code{\link[=setMultipleResources]{setMultipleResources()}}.
}
\keyword{internal}
