% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectra.R
\name{mr_plot_spectra_data}
\alias{mr_plot_spectra_data}
\title{Draw an assembled multiple-resource spectrum plot}
\usage{
mr_plot_spectra_data(
  df,
  params,
  y_label,
  size_axis,
  wlim,
  llim,
  ylim,
  log_x,
  log_y,
  highlight = NULL
)
}
\arguments{
\item{df}{The plotting data frame.}

\item{params}{A \linkS4class{mizerMR} object.}

\item{y_label}{The label for the y axis.}

\item{size_axis}{Either "w" or "l".}

\item{wlim, llim, ylim}{Limits.}

\item{log_x, log_y}{Whether the axes are logarithmic.}

\item{highlight}{Species to highlight.}
}
\value{
A ggplot object.
}
\description{
Draw an assembled multiple-resource spectrum plot
}
\keyword{internal}
