% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectra.R
\name{mr_resource_spectra_data}
\alias{mr_resource_spectra_data}
\title{Assemble the resource part of a spectrum plot}
\usage{
mr_resource_spectra_data(params, n_res, spectrum, size_axis, wlim, llim, ylim)
}
\arguments{
\item{params}{A \linkS4class{mizerMR} object.}

\item{n_res}{An array (resource x size) of resource number densities.}

\item{spectrum}{The list returned by \code{\link[=resolve_spectrum_power]{resolve_spectrum_power()}}.}

\item{size_axis}{Either "w" or "l".}

\item{wlim, llim, ylim}{Limits, with the weight limits already defaulted by
\code{\link[=mr_spectra_wlim]{mr_spectra_wlim()}}.}
}
\value{
A data frame with the variables 'w' (or 'l'), 'value', 'Spectra' and
'Legend'.
}
\description{
Turns an array of resource number densities into the same kind of plotting
data frame that mizer's \code{plotSpectra()} returns for the species, with the
size limits, the size axis and the value limits applied in the same order.
}
\keyword{internal}
