% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectra.R
\name{mr_spectra_wlim}
\alias{mr_spectra_wlim}
\title{Default weight limits for a multiple-resource spectrum plot}
\usage{
mr_spectra_wlim(params, wlim)
}
\arguments{
\item{params}{A \linkS4class{mizerMR} object.}

\item{wlim}{The weight limits supplied by the user.}
}
\value{
The weight limits with any NA replaced by the default.
}
\description{
The same defaults that mizer uses when the resource is shown: from a
hundredth of the smallest consumer size to the largest resource size.
}
\keyword{internal}
