% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerMR-package.R
\docType{class}
\name{MRMizerParams-class}
\alias{MRMizerParams-class}
\title{S4 subclass of MizerParams for models with multiple resources}
\description{
Created by \code{\link[=setMultipleResources]{setMultipleResources()}} or \code{\link[=newMRParams]{newMRParams()}}. Inherits all slots
from \linkS4class{MizerParams} and triggers S3 dispatch to mizerMR-specific
methods for plotting, accessors, and projection.
}
