% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{mrResourceView}
\alias{mrResourceView}
\title{Make a single-resource view of one resource}
\usage{
mrResourceView(
  base,
  resource_mort,
  abundance,
  resource_rate,
  resource_capacity
)
}
\arguments{
\item{base}{A plain \link[mizer:MizerParams-class]{mizer::MizerParams} object holding the model
without its extension classes.}

\item{resource_mort}{The mortality experienced by the resource.}

\item{abundance}{The number density of the resource.}

\item{resource_rate}{The replenishment rate of the resource.}

\item{resource_capacity}{The carrying capacity of the resource.}
}
\value{
A \code{mizerMRResourceView} object.
}
\description{
Make a single-resource view of one resource
}
\keyword{internal}
