% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NResource.R
\name{NResource.mizerMRSim}
\alias{NResource.mizerMRSim}
\title{Access resource abundances in simulation result}
\usage{
\method{NResource}{mizerMRSim}(sim)
}
\arguments{
\item{sim}{A MizerSim object}
}
\value{
For \code{NResource()}: An array (time x resource x size) holding the
resource number densities at all saved timesteps of the simulation.
}
\description{
Access resource abundances in simulation result
}
