% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NResource.R
\name{NResource}
\alias{NResource}
\alias{NResource.MRMizerSim}
\alias{finalNResource.MRMizerSim}
\title{Access resource abundances in simulation result}
\usage{
\method{NResource}{MRMizerSim}(sim, ...)

\method{finalNResource}{MRMizerSim}(sim, ...)
}
\arguments{
\item{sim}{A MizerSim object}

\item{...}{Unused}
}
\value{
For \code{NResource()}: An array (time x resource x size) holding the
resource number densities at all saved timesteps of the simulation.

For \code{finalNResource()}: An array (resource x size) holding the
resource number densities at the end of the simulation
}
\description{
Access resource abundances in simulation result
}
