% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet.R
\name{plotDietData}
\alias{plotDietData}
\title{Format and plot MR diet data}
\usage{
plotDietData(
  params,
  diet,
  species = NULL,
  wlim = c(NA, NA),
  llim = c(NA, NA),
  size_axis = c("w", "l"),
  log_x = TRUE,
  log_y = FALSE,
  log = NULL,
  return_data = FALSE
)
}
\arguments{
\item{params}{A \linkS4class{mizerMR} object.}

\item{diet}{A diet array as returned by \code{\link[=getDiet]{getDiet()}}.}

\item{species}{Optional predator species selection.}

\item{wlim, llim}{Limits on the predator size, on the weight and the length
axis respectively.}

\item{size_axis}{Whether to plot against predator weight ("w") or predator
length ("l").}

\item{log_x, log_y, log}{How the axes are scaled, see
\code{\link[mizer:parsePlotLog]{mizer::parsePlotLog()}}.}

\item{return_data}{Whether to return the plotting data instead of a plot.}
}
\value{
A ggplot2 object or, if \code{return_data = TRUE}, a data frame.
}
\description{
Converts an MR diet array to the data frame expected by \code{plotDataFrame()}, or
returns that data frame directly.
}
\keyword{internal}
