% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet.R
\name{plotDietMR}
\alias{plotDietMR}
\title{Plot diet, resolved by prey species, as function of predator at size.}
\usage{
plotDietMR(sim, species = NULL, time_range, xlim = c(1, NA), returnData = F)
}
\arguments{
\item{species}{The name of the predator species for which to plot the diet.}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the three variables 'w', 'Proportion', 'Prey' is returned.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Plots the proportions with which each prey species contributes to the total
biomass consumed by the specified predator species, as a function of the
predator's size. These proportions are obtained with \code{getDietMR()}.
}
\details{
Prey species that contribute less than 1 permille to the diet are suppressed
in the plot.
}
\seealso{
\code{\link[=getDietMR]{getDietMR()}}

Other plotting functions: 
\code{\link{animateSpectra}()},
\code{\link{plotSpectra}()}
}
\concept{plotting functions}
