% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResource.R
\name{plotResourceLevel}
\alias{plotResourceLevel}
\title{Plot the proportion of the resource spectrum(s) compared to
their carrying capacity}
\usage{
plotResourceLevel(object, return_data = FALSE)
}
\arguments{
\item{object}{An object of class \link[mizer:MizerSim-class]{mizer::MizerSim} or
\link[mizer:MizerParams-class]{mizer::MizerParams}.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is FALSE}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the three variables 'w', 'value', 'Resource' is returned.
}
\description{
The values are proportions, so as in mizer since version 3.3 they are shown
on a linear y axis covering the whole of the interval from 0 to 1, widened if
any resource is above its carrying capacity.
}
\examples{
\donttest{
plotResourceLevel(NS_params)

# Returning the data frame
fr <- plotResourceLevel(NS_params, return_data = TRUE)
str(fr)
}
}
\seealso{
\link[mizer:plotting_functions]{plotting_functions}

Other plotting functions:
\code{\link{animateSpectra}},
\code{\link{plotDiet}},
\code{\link[=plotResourcePred]{plotResourcePred()}},
\code{\link{plotSpectra}}
}
\concept{plotting functions}
