% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{project}
\alias{project}
\alias{project.MRMizerParams}
\title{Project an MRMizerParams model forward in time}
\usage{
\method{project}{MRMizerParams}(params, ...)
}
\arguments{
\item{params}{An \linkS4class{MRMizerParams} object.}

\item{...}{Further arguments passed to \code{\link[mizer:project]{mizer::project()}}.}
}
\value{
An object of class \linkS4class{MRMizerSim}.
}
\description{
S3 method for \code{\link[=project]{project()}} when called with an \linkS4class{MRMizerParams}
object. Delegates all work to mizer's standard \code{project()} and wraps the
result as an \linkS4class{MRMizerSim} so that mizerMR's accessor and
plotting methods are dispatched automatically on the simulation output.
}
