% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R
\name{project_methods}
\alias{project_methods}
\alias{mizerMRBaseResource}
\alias{mizerMRValidBaseResource}
\alias{getEncounter.mizerMR}
\alias{getPredRate.mizerMR}
\alias{getResourceMort.mizerMR}
\alias{getRates.mizerMR}
\alias{projectEncounter.mizerMR}
\alias{mizerMREncounter}
\alias{mizerMRResourceEncounter}
\alias{projectResourceMort.mizerMR}
\title{Multiple-resource project methods}
\usage{
mizerMRBaseResource(params)

mizerMRValidBaseResource(params, n_pp)

\method{getEncounter}{mizerMR}(
  object,
  n = initialN(object),
  n_pp = mizerMRBaseResource(object),
  n_other = initialNOther(object),
  t = 0,
  ...
)

\method{getPredRate}{mizerMR}(
  object,
  n = initialN(object),
  n_pp = mizerMRBaseResource(object),
  n_other = initialNOther(object),
  t = 0,
  ...
)

\method{getResourceMort}{mizerMR}(
  params,
  n = initialN(params),
  n_pp = mizerMRBaseResource(params),
  n_other = initialNOther(params),
  t = 0,
  ...
)

\method{getRates}{mizerMR}(
  params,
  n = initialN(params),
  n_pp = mizerMRBaseResource(params),
  n_other = initialNOther(params),
  effort,
  t = 0,
  ...
)

\method{projectEncounter}{mizerMR}(params, n, n_pp, n_other, t = 0, ...)

mizerMREncounter(params, n, n_pp, n_other, t = 0, ...)

mizerMRResourceEncounter(params, n_other)

\method{projectResourceMort}{mizerMR}(params, n, n_pp, n_other, t = 0, pred_rate, ...)
}
\arguments{
\item{params}{A \linkS4class{mizerMR} object.}

\item{n_pp}{A resource abundance vector, or an MR resource array that will be
replaced by the silenced built-in resource vector before calling mizer's
base method.}

\item{n}{A matrix of species abundances (species x size).}

\item{n_other}{A list of abundances for other dynamical components.}

\item{t}{Current projection time.}

\item{...}{Further arguments passed along the mizer method chain.}

\item{effort}{Fishing effort by gear.}
}
\value{
The return value depends on the method called: a rate array, a rate
list, or a resource abundance vector.
}
\description{
Internal helpers and S3 methods used to make mizer's rate and projection
generics work with the multiple-resource component.

Internal projection hook for predation mortality on multiple resources.
}
\keyword{internal}
