% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_utils.R
\name{removeSpecies}
\alias{removeSpecies}
\alias{removeSpecies.mizerMR}
\title{Remove species from a mizerMR model}
\usage{
\method{removeSpecies}{mizerMR}(params, species, ...)
}
\arguments{
\item{params}{A \linkS4class{mizerMR} object.}

\item{species}{The species to be removed. A vector of species names, or a
numeric vector of species indices, or a logical vector indicating for
each species whether it is to be removed (TRUE) or not.}

\item{...}{Currently unused.}
}
\value{
A \linkS4class{mizerMR} object with the specified species removed.
}
\description{
Extends \code{\link[mizer:removeSpecies]{mizer::removeSpecies()}} to also trim the resource interaction
matrix.
}
