% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_utils.R
\name{renameSpecies}
\alias{renameSpecies}
\alias{renameSpecies.MRMizerParams}
\title{Rename species in an MRMizerParams model}
\usage{
\method{renameSpecies}{MRMizerParams}(params, replace)
}
\arguments{
\item{params}{An \linkS4class{MRMizerParams} object.}

\item{replace}{A named character vector, with new names as values, and old
names as names.}
}
\value{
An \linkS4class{MRMizerParams} object with the species renamed.
}
\description{
Extends \code{\link[mizer:renameSpecies]{mizer::renameSpecies()}} to also update the row names of the
resource interaction matrix.
}
