% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SEAmodel}
\alias{SEAmodel}
\title{South-East Australian reef ecosystem model}
\format{
A \link[=mizerMR-class]{mizerMR} object (an S4 object extending
\link[mizer:MizerParams-class]{mizer::MizerParams}) with 17 species and 3 resources.
The multiple resources are stored in \code{params@other_params[["MR"]]}.
}
\source{
Created with \code{\link[=setMultipleResources]{setMultipleResources()}}.
Original parameter derivations: \url{https://github.com/astaaudzi/SEAmodel}.
}
\usage{
data(SEAmodel)
}
\description{
A mizerMR model object representing a size-structured multispecies
model of the South-East Australian (SEA) rocky reef coast ecosystem.
The model contains 17 species and three size-structured background resource
spectra representing distinct primary production pathways:
\describe{
\item{pl}{Pelagic plankton resource}
\item{aa}{Algal/attached benthic resource}
\item{bb}{Benthic invertebrate resource}
}
}
\details{
The original model code and parameter derivations are available at
\url{https://github.com/astaaudzi/SEAmodel}.
}
\references{
Audzijonyte A, Delius GW, Stuart-Smith RD, Novaglio C, Edgar GJ,
Barrett NS, Blanchard JL (2023). Changes in sea floor productivity are
crucial to understanding the impact of climate change in temperate coastal
ecosystems according to a new size-based model.
\emph{PLoS Biology} 21(12): e3002392.
\doi{10.1371/journal.pbio.3002392}
}
\seealso{
\code{\link[=setMultipleResources]{setMultipleResources()}}, \code{\link[=newMRParams]{newMRParams()}}
}
\keyword{datasets}
