% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_params.R
\name{set_resource_param_default}
\alias{set_resource_param_default}
\title{Set a resource parameter to a default value}
\usage{
set_resource_param_default(object, parname, default, message = NULL)
}
\arguments{
\item{object}{Either a MizerParams object or a resource parameter data frame}

\item{parname}{A string with the name of the resource parameter to set}

\item{default}{A single default value or a vector with one default value for
each resource}

\item{message}{A string with a message to be issued when the parameter did
not already exist}
}
\value{
The \code{object} with an updated column in the resource params data frame.
}
\description{
If the resource parameter does not yet exist in the resource parameter data
frame, then create it and fill it with the default. Otherwise use the default
only to fill in any NAs. Optionally gives a message if the parameter
did not already exist.
}
\concept{helper}
