% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMultipleResources.R
\name{setMultipleResources}
\alias{setMultipleResources}
\alias{setMultipleResourcesInternal}
\alias{resource_capacity}
\alias{resource_capacity<-}
\alias{resource_rate}
\alias{resource_rate<-}
\alias{resource_interaction}
\alias{resource_interaction<-}
\alias{initialNResource.mizerMR}
\alias{initialNResource<-.mizerMR}
\title{Set up multiple resources}
\usage{
setMultipleResources(
  params,
  resource_params = NULL,
  resource_interaction = NULL,
  resource_capacity = NULL,
  resource_rate = NULL,
  initial_resource = NULL,
  info_level = mizer::default_info_level()
)

setMultipleResourcesInternal(
  params,
  resource_params = NULL,
  resource_interaction = NULL,
  resource_capacity = NULL,
  resource_rate = NULL,
  initial_resource = NULL,
  info_level = mizer::default_info_level()
)

resource_capacity(params)

resource_capacity(params) <- value

resource_rate(params)

resource_rate(params) <- value

resource_interaction(params)

resource_interaction(params) <- value

\method{initialNResource}{mizerMR}(object)

\method{initialNResource}{mizerMR}(params) <- value
}
\arguments{
\item{params}{A MizerParams object}

\item{resource_params}{A data frame with the resource parameters}

\item{resource_interaction}{Optional interaction matrix between species and
resources (predator species x prey resource). By default all entries are 1.}

\item{resource_capacity}{Optional. Array (resource x size) of the
intrinsic resource carrying capacities}

\item{resource_rate}{Optional. Array (resource x size) of intrinsic
resource growth rates}

\item{initial_resource}{Optional. Array (resource x size) of initial values}

\item{info_level}{Controls the amount of information the function reports
about the choices it makes, see \code{\link[mizer:default_info_level]{mizer::default_info_level()}}.}

\item{value}{Value to assign}

\item{object}{A mizerMR object}
}
\description{
Set up multiple resources
}
\keyword{internal}
