% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade.mizerMR}
\alias{upgrade.mizerMR}
\title{Upgrade a mizerMR object from an earlier version of mizerMR}
\usage{
\method{upgrade}{mizerMR}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{mizerMR} object to be upgraded.}

\item{...}{Unused.}
}
\value{
The upgraded object.
}
\description{
This is the \code{mizerMR} method of mizer's \code{\link[mizer:upgrade]{mizer::upgrade()}} generic. It
performs only the mizerMR-specific migration and is invoked by mizer's
upgrade orchestrator (from \code{validParams()} / \code{readParams()}) when the version
stamp recorded in \code{params@extensions[["mizerMR"]]} is older than the installed
mizerMR, or is missing. The orchestrator records the new stamp afterwards, so
this method must \strong{not} stamp the version itself and must \strong{not} call
\code{NextMethod()}. It is written to be idempotent.
}
\details{
Migrations performed:
\itemize{
\item The \code{resource_params} data frame used to be stored under
\code{other_params(params)[["MR"]]$resource_params}, i.e. in
\code{params@other_params$other$MR$resource_params}. It now lives alongside the
other MR component parameters in \code{params@other_params$MR$resource_params}.
This method moves it to the new location.
}
}
\keyword{internal}
