% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMultipleResources.R
\name{valid_initial_resource}
\alias{valid_initial_resource}
\title{Return valid initial resource array}
\usage{
valid_initial_resource(params, initial_resource = NULL)
}
\arguments{
\item{params}{A MizerParams object}

\item{initial_resource}{Array (resource x size) of initial values}
}
\value{
An array (resource x size)
}
\description{
If \code{initial_resource} is given it is checked for validity and returned.
Otherwise the value stored in \code{params} is returned.
}
