% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMultipleResources.R
\name{valid_resource_capacity}
\alias{valid_resource_capacity}
\title{Return valid resource capacity array}
\usage{
valid_resource_capacity(
  params,
  resource_params = NULL,
  resource_capacity = NULL
)
}
\arguments{
\item{params}{A MizerParams object}

\item{resource_capacity}{Array (resource x size) of the
intrinsic resource carrying capacities}
}
\value{
An array (resource x size) with the resource capacities
}
\description{
If \verb{resource capacity} is given it is checked for validity. If it does not
have a comment, then it is given the comment "set manually". This is then
returned. If \verb{resource capacity} is missing or NULL, but one was set by the
user and stored in \code{params} with a comment, then this is returned. Otherwise
a resource capacity is calculated from \code{resource_params}. If this is NULL to
it is taken from \code{params}.
}
