% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMultipleResources.R
\name{valid_resource_interaction}
\alias{valid_resource_interaction}
\title{Return valid resource interaction array}
\usage{
valid_resource_interaction(params, resource_interaction = NULL)
}
\arguments{
\item{params}{A MizerParams object}

\item{resource_interaction}{Interaction matrix between species and
resources (predator species x prey resource). By default all entries are 1.}
}
\value{
An array (resource x size)
}
\description{
If \verb{resource interaction} is given it is checked for validity and returned.
Otherwise the value stored in \code{params} is returned.
}
