% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_params.R
\name{validResourceParams}
\alias{validResourceParams}
\title{Validate resource parameter data frame}
\usage{
validResourceParams(resource_params, params)
}
\arguments{
\item{resource_params}{The user-supplied resource parameter data frame}

\item{min_w}{The smallest allowed resource size}
}
\value{
A valid resource parameter data frame

This function throws an error if
\itemize{
\item the \code{resource} column does not exist or contains duplicates
\item \code{w_min} is not smaller than \code{w_max}
\item \code{w_min} is smaller than \code{min_w}
\item any parameter is negative
}

Any missing values are set to the values that were previously set for the
first resource (which would be the standard single resource when you set up
multiple resources for the first time). The exception is the \code{colour} which
of course should be different for each resource. It is chosen from a
colour-blind-friendly palette.

Where no values are available for the previous first resource either,
the following defaults are used:
\itemize{
\item \code{kappa} is set to \code{0.1}
\item \code{lambda} is set to \code{2.05}
\item \code{r_pp} is set to \code{4}
\item \code{w_min} is set to \code{min_w}
\item \code{w_max} is set to \code{10}
\item \code{n} is set to 2/3
\item \code{dynamics} is set to \code{resource_semichemostat}
\item \code{linetype} is set to "solid"
}

If \code{resource_params} was provided as a tibble it is converted back to an
ordinary data frame.
}
\description{
Check validity of resource parameters and set defaults for missing but
required parameters
}
\concept{helper}
