% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerMR-class.R
\name{mizerMR-class}
\alias{mizerMR-class}
\alias{mizerMRSim-class}
\title{mizerMR marker classes}
\description{
S4 marker subclasses of \link[mizer:MizerParams]{MizerParams} and \link[mizer:MizerSim]{MizerSim} that enable S3 dispatch
for MR-specific methods.
}
\details{
Objects of class \code{mizerMR} are created by \code{\link[=setMultipleResources]{setMultipleResources()}}.
Objects of class \code{mizerMRSim} are returned automatically by \code{\link[mizer:project]{project()}}
when called on a \code{mizerMR} params object.
}
