% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerMR-package.R
\docType{package}
\name{mizerMR-package}
\alias{mizerMR}
\alias{mizerMR-package}
\title{mizerMR: Add multiple size-structured resources to a mizer model}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An add-on to the mizer package which allows multiple size-structured resource spectra to be added to any mizer model.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://sizespectrum.org/mizerMR/}
  \item \url{https://github.com/sizespectrum/mizerMR}
  \item Report bugs at \url{https://github.com/sizespectrum/mizerMR/issues}
}

}
\author{
\strong{Maintainer}: Gustav Delius \email{gustav.delius@york.ac.uk} (\href{https://orcid.org/0000-0003-4092-8228}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Romain Forestier \email{romain.forestier@utas.edu.au}
}

}
\keyword{internal}
