% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMRParams.R
\name{newMRparams}
\alias{newMRparams}
\title{Create a new multi-resource, multi-species model}
\usage{
newMRparams(
  species_params,
  resource_params,
  gear_params,
  interaction,
  interaction_resource
)
}
\arguments{
\item{interaction}{Optional interaction matrix between species (predator
species x prey species). Entries should be numbers between 0 and 1. By
default all entries are 1.}

\item{interation_resource}{Optional interaction matrix between species and
resources (predator species x prey resource). Entries should be numbers
between 0 and 1. By default all entries are 1.}
}
\value{
An object of class MizerMRParams
}
\description{
Create a new multi-resource, multi-species model
}
