% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NResource.R
\name{NResource}
\alias{NResource}
\alias{finalNResource}
\title{Access resource abundances in simulation result}
\usage{
NResource(sim)

finalNResource(sim)
}
\value{
For \code{finalNResource()}: An array (resource x size) holding the
resource number densities at the end of the simulation
}
\description{
Access resource abundances in simulation result
}
