% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayResource-class.R
\name{plot.MRArrayResourceBySize}
\alias{plot.MRArrayResourceBySize}
\alias{plotHover.MRArrayResourceBySize}
\alias{plot.MRArrayTimeByResourceBySize}
\alias{plotHover.MRArrayTimeByResourceBySize}
\title{Plot mizerMR resource arrays}
\usage{
\method{plot}{MRArrayResourceBySize}(
  x,
  resources = NULL,
  return_data = FALSE,
  log_x = TRUE,
  log_y = TRUE,
  log = NULL,
  wlim = c(NA, NA),
  ylim = c(NA, NA),
  y_ticks = 6,
  ...
)

\method{plotHover}{MRArrayResourceBySize}(x, ...)

\method{plot}{MRArrayTimeByResourceBySize}(x, time = NULL, ...)

\method{plotHover}{MRArrayTimeByResourceBySize}(x, ...)
}
\arguments{
\item{x}{An \code{MRArrayResourceBySize} or \code{MRArrayTimeByResourceBySize} object.}

\item{resources}{Optional vector of resource names (or indices) to restrict
the plot to.}

\item{return_data}{Logical. If TRUE, return the plotting data frame instead
of the plot.}

\item{log_x, log_y}{Logical flags for logarithmic axes.}

\item{log}{Alternative way to specify the log axes, see
\code{\link[mizer:parsePlotLog]{mizer::parsePlotLog()}}.}

\item{wlim, ylim}{Length-2 numeric vectors giving the weight and value limits.}

\item{y_ticks}{Approximate number of ticks on the y-axis.}

\item{...}{Unused.}

\item{time}{For \code{MRArrayTimeByResourceBySize}, the time at which to plot the
spectrum. Defaults to the final time step.}
}
\value{
A ggplot object, or a data frame if \code{return_data = TRUE}.
}
\description{
\code{plot()} creates a ggplot2 figure with one line for each resource, showing
the values against size. It is the multiple-resource analogue of the
\code{plot()} method for mizer's \code{ArrayResourceBySize} objects and reuses the same
internal mizer plotting machinery, so resources are drawn in the colours
registered in the \code{params} object.
}
