% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet.R
\name{plotDiet}
\alias{plotDiet}
\alias{plotDiet.mizerMR}
\alias{plotDiet.mizerMRSim}
\alias{plotDietMR}
\title{Plot diet, resolved by prey species, as function of predator at size.}
\usage{
\method{plotDiet}{mizerMR}(
  object,
  species = NULL,
  wlim = c(1, NA),
  llim = c(NA, NA),
  size_axis = c("w", "l"),
  return_data = FALSE,
  log_x = TRUE,
  log_y = FALSE,
  log = NULL,
  ...,
  time_range
)

\method{plotDiet}{mizerMRSim}(
  object,
  species = NULL,
  wlim = c(1, NA),
  llim = c(NA, NA),
  size_axis = c("w", "l"),
  return_data = FALSE,
  log_x = TRUE,
  log_y = FALSE,
  log = NULL,
  ...,
  time_range
)

plotDietMR(object, ...)
}
\arguments{
\item{object}{An object of class \link[mizer:MizerSim-class]{mizer::MizerSim} or
\link[mizer:MizerParams-class]{mizer::MizerParams}.}

\item{species}{The name of the predator species for which to plot the diet.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the w axis. Use NA to refer to the existing minimum or maximum.}

\item{llim}{Ignored (compatibility argument from mizer's generic).}

\item{size_axis}{Ignored (compatibility argument from mizer's generic).}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is FALSE.}

\item{log_x}{Ignored (compatibility argument from mizer's generic).}

\item{log_y}{Ignored (compatibility argument from mizer's generic).}

\item{log}{Ignored (compatibility argument from mizer's generic).}

\item{...}{Other arguments (currently unused).}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step. Only used when called with a \link[mizer:MizerSim-class]{mizer::MizerSim}
object.}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the three variables 'w', 'Proportion', 'Prey' is returned.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Plots the proportions with which each prey species contributes to the total
biomass consumed by the specified predator species, as a function of the
predator's size. These proportions are obtained with \code{getDiet()}.
}
\details{
Prey species that contribute less than 1 permille to the diet are suppressed
in the plot.
}
\seealso{
\code{\link[=getDiet]{getDiet()}}

Other plotting functions:
\code{\link{animateSpectra}},
\code{\link[=plotResourceLevel]{plotResourceLevel()}},
\code{\link[=plotResourcePred]{plotResourcePred()}},
\code{\link{plotSpectra}}
}
\concept{plotting functions}
