% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet.R
\name{plotDietData}
\alias{plotDietData}
\title{Format and plot MR diet data}
\usage{
plotDietData(params, diet, species = NULL, return_data = FALSE)
}
\arguments{
\item{params}{A \linkS4class{mizerMR} object.}

\item{diet}{A diet array as returned by \code{\link[=getDiet]{getDiet()}}.}

\item{species}{Optional predator species selection.}

\item{return_data}{Whether to return the plotting data instead of a plot.}
}
\value{
A ggplot2 object or, if \code{return_data = TRUE}, a data frame.
}
\description{
Converts an MR diet array to the data frame expected by \code{plotDataFrame()}, or
returns that data frame directly.
}
\keyword{internal}
