% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet.R
\name{plotDietMR}
\alias{plotDietMR}
\title{Plot diet, resolved by prey species, as function of predator at size.}
\usage{
plotDietMR(
  object,
  species = NULL,
  time_range,
  wlim = c(1, NA),
  return_data = FALSE
)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim} or
\linkS4class{MizerParams}.}

\item{species}{The name of the predator species for which to plot the diet.}

\item{wlim}{A numeric vector of length two providing lower and upper limits
for the w axis. Use NA to refer to the existing minimum or maximum.}

\item{return_data}{A boolean value that determines whether the formatted data
used for the plot is returned instead of the plot itself. Default value is FALSE}
}
\value{
A ggplot2 object, unless \code{return_data = TRUE}, in which case a data
frame with the three variables 'w', 'Proportion', 'Prey' is returned.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Plots the proportions with which each prey species contributes to the total
biomass consumed by the specified predator species, as a function of the
predator's size. These proportions are obtained with \code{getDietMR()}.
}
\details{
Prey species that contribute less than 1 permille to the diet are suppressed
in the plot.
}
\seealso{
\code{\link[=getDietMR]{getDietMR()}}

Other plotting functions: 
\code{\link{animateSpectra}()},
\code{\link{plotSpectra}()}
}
\concept{plotting functions}
