% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_utils.R
\name{removeSpecies}
\alias{removeSpecies}
\alias{removeSpecies.MRMizerParams}
\title{Remove species from an MRMizerParams model}
\usage{
\method{removeSpecies}{MRMizerParams}(params, species)
}
\arguments{
\item{params}{An \linkS4class{MRMizerParams} object.}

\item{species}{The species to be removed. A vector of species names, or a
numeric vector of species indices, or a logical vector indicating for
each species whether it is to be removed (TRUE) or not.}
}
\value{
An \linkS4class{MRMizerParams} object with the specified species removed.
}
\description{
Extends \code{\link[mizer:removeSpecies]{mizer::removeSpecies()}} to also trim the resource interaction
matrix.
}
