% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_utils.R
\name{renameSpecies}
\alias{renameSpecies}
\alias{renameSpecies.mizerMR}
\title{Rename species in a mizerMR model}
\usage{
\method{renameSpecies}{mizerMR}(params, replace, ...)
}
\arguments{
\item{params}{A \linkS4class{mizerMR} object.}

\item{replace}{A named character vector, with new names as values, and old
names as names.}

\item{...}{Currently unused.}
}
\value{
A \linkS4class{mizerMR} object with the species renamed.
}
\description{
Extends \code{\link[mizer:renameSpecies]{mizer::renameSpecies()}} to also update the row names of the
resource interaction matrix.
}
