% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_params.R
\name{resource_params}
\alias{resource_params}
\alias{resource_params<-}
\title{Resource parameters}
\usage{
resource_params(params)

resource_params(params) <- value
}
\arguments{
\item{params}{A MizerParams object}

\item{value}{A data frame with the resource parameters}
}
\description{
The resource parameter data frame has one row for each resource. The
variables are:
\itemize{
\item \code{resource} Name you want to use for the resource
\item \code{kappa}    Coefficient in the carrying capacity power law
\item \code{lambda}   Exponent of the carrying capacity power law
\item \code{r_pp}     Coefficient in the allometric replenishment rate
\item \code{w_min}    Smallest size of the resource
\item \code{w_max}    Largest size of the resource
\item \code{resource_dynamics} Name of the resource dynamics function
}
}
\seealso{
\code{\link[=validResourceParams]{validResourceParams()}}
}
\concept{functions for setting parameters}
