% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_methods.R, R/resource_dynamics.R,
%   R/setMultipleResources.R
\name{getEncounter.MRMizerParams}
\alias{getEncounter.MRMizerParams}
\alias{getResourceMort.MRMizerParams}
\alias{setMultipleResources}
\alias{resource_capacity.MRMizerParams}
\alias{resource_capacity<-.MRMizerParams}
\alias{resource_rate.MRMizerParams}
\alias{resource_rate<-.MRMizerParams}
\alias{resource_interaction}
\alias{resource_interaction.MRMizerParams}
\alias{resource_interaction<-}
\alias{resource_interaction<-.MRMizerParams}
\alias{initialNResource.MRMizerParams}
\alias{initialNResource<-.MRMizerParams}
\title{Set up multiple resources}
\usage{
\method{getEncounter}{MRMizerParams}(
  params,
  n = initialN(params),
  n_pp = params@initial_n_pp,
  n_other = initialNOther(params),
  t = 0,
  ...
)

\method{getResourceMort}{MRMizerParams}(params, ...)

setMultipleResources(
  params,
  resource_params = NULL,
  resource_interaction = NULL,
  resource_capacity = NULL,
  resource_rate = NULL,
  initial_resource = NULL
)

\method{resource_capacity}{MRMizerParams}(params)

\method{resource_capacity}{MRMizerParams}(params) <- value

\method{resource_rate}{MRMizerParams}(params)

\method{resource_rate}{MRMizerParams}(params) <- value

resource_interaction(params, ...)

\method{resource_interaction}{MRMizerParams}(params, ...)

resource_interaction(params) <- value

\method{resource_interaction}{MRMizerParams}(params) <- value

\method{initialNResource}{MRMizerParams}(params)

\method{initialNResource}{MRMizerParams}(params) <- value
}
\arguments{
\item{params}{A MizerParams object}

\item{resource_params}{A data frame with the resource parameters}

\item{resource_interaction}{Optional interaction matrix between species and
resources (predator species x prey resource). By default all entries are 1.}

\item{resource_capacity}{Optional. Array (resource x size) of the
intrinsic resource carrying capacities}

\item{resource_rate}{Optional. Array (resource x size) of intrinsic
resource growth rates}

\item{initial_resource}{Optional. Array (resource x size) of initial values}

\item{value}{Value to assign}
}
\description{
Set up multiple resources
}
