% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMultipleResources.R
\name{setMultipleResources}
\alias{setMultipleResources}
\alias{resource_capacity}
\alias{resource_capacity<-}
\alias{resource_rate}
\alias{resource_rate<-}
\alias{resource_interaction}
\alias{resource_interaction<-}
\alias{initialNResource}
\alias{initialNResource<-}
\title{Set up multiple resources}
\usage{
setMultipleResources(
  params,
  resource_params = params@resource_params,
  resource_interaction = NULL,
  resource_capacity = NULL,
  resource_rate = NULL,
  initial_resource = resource_capacity
)

resource_capacity(params)

resource_capacity(params) <- value

resource_rate(params)

resource_rate(params) <- value

resource_interaction(params)

resource_interaction(params) <- value

initialNResource(params)

initialNResource(params) <- value
}
\arguments{
\item{params}{A MizerParams object}

\item{resource_params}{A data frame with the resource parameters}

\item{resource_interaction}{Optional interaction matrix between species and
resources (predator species x prey resource). By default all entries are 1.}

\item{resource_capacity}{Optional. Array (resource x size) of the
intrinsic resource carrying capacities}

\item{resource_rate}{Optional. Array (resource x size) of intrinsic
resource growth rates}

\item{initial_resource}{Optional. Array (resource x size) of initial values}

\item{value}{Value to assign}
}
\description{
Set up multiple resources
}
