% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newMRParams.R
\name{setMultipleResources}
\alias{setMultipleResources}
\title{Set up multiple resources}
\usage{
setMultipleResources(
  params,
  resource_params = params@resource_params,
  resource_interaction = NULL,
  resource_capacity = NULL,
  resource_rate = NULL,
  initial_resource = resource_capacity
)
}
\arguments{
\item{params}{A MizerParams object}

\item{resource_params}{A data frame with the resource parameters}

\item{resource_interaction}{Optional interaction matrix between species and
resources (predator species x prey resource). Entries should be numbers
between 0 and 1. By default all entries are 1.}

\item{resource_capacity}{Optional. Array (resource x size) of the
intrinsic resource carrying capacities}

\item{resource_rate}{Optional. Array (resource x size) of intrinsic
resource growth rates}

\item{initial_resource}{Optional. Array (resource x size) of initial values}
}
\description{
Set up multiple resources
}
