% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_params.R
\name{validResourceParams}
\alias{validResourceParams}
\title{Validate resource parameter data frame}
\usage{
validResourceParams(resource_params, min_w)
}
\arguments{
\item{resource_params}{The user-supplied resource parameter data frame}

\item{min_w}{The smallest allowed resource size}
}
\value{
A valid resource parameter data frame

This function throws an error if
\itemize{
\item the \code{resource} column does not exist or contains duplicates
\item \code{w_min} is not smaller than \code{w_max}
\item \code{w_min} is smaller than \code{min_w}
\item any parameter is negative
}

It sets default values if any of the following are missing or NA
\itemize{
\item \code{kappa} is set to \code{0.1}
\item \code{lambda} is set to \code{2.05}
\item \code{r_pp} is set to \code{4}
\item \code{w_min} is set to \code{min_w}
\item \code{w_max} is set to \code{10}
\item \code{dynamics} is set to \code{semichemostat}
\item \code{colour} is drawn from a colour-blind-friendly palette
\item \code{linetype} is set to "solid"
}

If \code{resource_params} was provided as a tibble it is converted back to an
ordinary data frame.
}
\description{
Check validity of resource parameters and set defaults for missing but
required parameters
}
\concept{helper}
